+++ massiv/src/Data/Massiv/Array/Unsafe.hs
@@ -61,6 +61,8 @@ module Data.Massiv.Array.Unsafe
   , unsafeAtomicOrIntArray
   , unsafeAtomicXorIntArray
   , unsafeCasIntArray
+    -- ** Other operations
+  , unsafeUnstablePartitionRegionM
   ) where
 
 import Data.Massiv.Array.Delayed.Pull (D)
@@ -70,6 +72,7 @@ import Data.Massiv.Array.Manifest.Storable
 import Data.Massiv.Core.Common
 import Data.Massiv.Core.Index.Internal (Sz(SafeSz))
 import Data.Massiv.Core.Index.Stride (Stride(SafeStride))
+import Data.Massiv.Array.Ops.Sort (unsafeUnstablePartitionRegionM)
 
 unsafeBackpermute :: (Source r' ix' e, Index ix) =>
                      Sz ix -> (ix -> ix') -> Array r' ix' e -> Array D ix e